rgw: Changes for issue #14934 - rgw related changes only

+++ src/rgw/rgw_op.cc
@@ -3970,7 +3970,7 @@ void RGWCompleteMultipart::execute()
     return;
   }
 
-  if ((int)parts->parts.size() >
+  if (parts->parts.size() >
       s->cct->_conf->rgw_multipart_part_upload_limit) {
     op_ret = -ERANGE;
     return;
